Move provider & test provider to their own files
Move the tests as well

Rename `prettier` task to `format`
