+++ .husky/pre-commit
@@ -1,1 +1,1 @@
-npm run lint && npm run prettier && npm run test && npm run build
+npm run lint && npm run format && npm run test && npm run build
