+++ package.json
@@ -21,8 +21,8 @@
     "pretest": "node scripts/update-version.js",
     "test": "tsup && jest --verbose",
     "lint": "eslint --ext .ts,.tsx src/",
-    "prettier": "prettier . -l",
-    "prettier:fix": "npx prettier --write .",
+    "format": "prettier . -l",
+    "format:fix": "npx prettier --write .",
     "release": "npm run prettier:fix && npm run lint && npm run test && npm run build",
     "postversion": "git push && git push --tags",
     "version": "echo 'version' && npm run release && npm run prettier:fix && npm run build && git add src/version.ts",
